+++ src/ambition/version_details.cc
@@ -17,6 +17,13 @@ std::vector<VersionUpdates> versionDetails = {
   {
     .identifier = IDENTIFIER_PREVIEW,
     .sections = {
+      {
+        .heading = HEADING_FIXED,
+        .items = {
+          _("Used translated War Machine names on the War Factory building information panels."),
+          _("Used translated Greater Being names on the Seat of Power building information panels."),
+        },
+      },
     },
   },
   {