+++ CHANGELOG.md
@@ -10,6 +10,13 @@ speaking, considered to be the player's experience.
 
 ## [Unreleased]
 
+### Fixed
+
+- Used translated War Machine names on the War Factory building information
+  panels.
+- Used translated Greater Being names on the Seat of Power building information
+  panels.
+
 
 ## [3.2.9] — 2026-06-05